+++ flow/__init__.py
@@ -2,9 +2,9 @@ from __future__ import annotations
 
 __version__ = "0.0.1"
 
-from flow.agent import Agent, RunResult
-from flow.model import ChatResponse, Model, ToolCall
-from flow.tool import Tool, build_schema, tool
+from flow.lib.agent import Agent, RunResult
+from flow.lib.model import ChatResponse, Model, ToolCall
+from flow.lib.tool import Tool, build_schema, tool
 
 __all__ = [
 	"Agent",
