+++ flow/ai/doctype/ai_agent/test_ai_agent.py
@@ -7,8 +7,8 @@ from unittest.mock import patch
 import frappe
 from frappe.tests import IntegrationTestCase
 
-from flow.agent import Agent
-from flow.model import ChatResponse, Model
+from flow.lib.agent import Agent
+from flow.lib.model import ChatResponse, Model
 from flow.tools.builtins import sync_builtin_tools
